+++ examples/supply_chain/rl/__init__.py
@@ -1,9 +1,9 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT license.
 
-from .common.callbacks import post_collect, post_evaluate
-from .rl.env_sampler import env_sampler_creator
-from .rl.policy_registry import agent2policy, policy_creator, trainable_policies, trainer_creator
+from .callbacks import post_collect, post_evaluate
+from .env_sampler import env_sampler_creator
+from .policy_registry import agent2policy, policy_creator, trainable_policies, trainer_creator
 
 __all__ = [
     "agent2policy",
