+++ maro/rl/training/algorithms/base/ac_ppo_base.py
@@ -276,6 +276,7 @@ class DiscreteACBasedTrainer(SingleAgentTrainer):
 
     def train_step(self) -> None:
         assert isinstance(self._ops, DiscreteACBasedOps)
+        print(f"{self._name} training...")
         batch = self._get_batch()
         for _ in range(self._params.grad_iters):
             self._ops.update_critic(batch)
