+++ maro/rl/workflows/main.py
@@ -122,3 +122,9 @@ def main(scenario: Scenario) -> None:
     if isinstance(env_sampler, BatchEnvSampler):
         env_sampler.exit()
     training_manager.exit()
+
+
+if __name__ == "__main__":
+    # get user-defined scenario ingredients
+    scenario = Scenario(get_env("SCENARIO_PATH"))
+    main(scenario)
